Prevent the ball from getting stuck in the paddle
Previously it was possible to get the ball "stuck" in the paddle where
it would bounce back and forth until it eventually freed itself. This
change prevents that from happening. If this causes problems or you want
to disable it, you can change the new `USE_BALL_STUCK_IN_PADDLE_FIX`
config setting in config.py.